backend/pyproject.toml
[tool.poetry]
name = "taskflow-backend"
version = "0.1.0"
description = "Backend API for TaskFlow - task management system"
authors = ["SonderLor <sonder1or@yandex.ru>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
# FastAPI и зависимости
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

# База данных
sqlalchemy = "^2.0.23"
asyncpg = "^0.28.0"
greenlet = "^3.0.1"
alembic = "^1.12.1"

# Веб-сокеты
websockets = "^11.0.3"

# Утилиты
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"
email-validator = "^2.1.0.post1"

# CORS
starlette = "^0.27.0"

[tool.poetry.group.dev.dependencies]
# Тестирование
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
skip_glob = "__init__.py"
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
exclude = ["**/__init__.py"]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
